+++ src/Physics/Shapes/ReactTerrainShape.cpp
@@ -2,7 +2,7 @@
 // Created by Charlie Sewell on 13/05/2021.
 //
 
-#include <Components/Transform.hpp>
+#include "Transform.hpp"
 #include "ReactTerrainShape.hpp"
 
 ReactTerrainShape::ReactTerrainShape()
@@ -17,11 +17,11 @@ ReactTerrainShape::~ReactTerrainShape()
 void ReactTerrainShape::CreateTerrainShape(reactphysics3d::PhysicsCommon &physicsCommon, std::vector<std::vector<float>> heightVals)
 {
     m_heightFieldArray = new float[100*100];
-    for(int x=0; x<100; ++x)
+    for(size_t x=0; x<100; ++x)
     {
-        for(int z=0; z<100; ++z)
+        for(size_t z=0; z<100; ++z)
         {
-            m_heightFieldArray[x * 100 + z] = heightVals.at(static_cast<float>(z)).at(static_cast<float>(x));
+            m_heightFieldArray[x * 100 + z] = heightVals.at(z).at(x);
             //std::cout << heightFieldArray[x * 100 + z] <<std::endl;
         }
     }
